fix(38-01): complex_gap_stats PERCENTILE_CONT에 ::numeric 추가 — 프로덕션 정의와 일치
저장소 파일이 프로덕션 함수 정의와 달랐다(파일↔프로덕션 drift).
transactions.price가 bigint라 PERCENTILE_CONT는 double precision을 반환하고,
캐스트 없이는 116·117행의 ROUND(<double precision>, 1)이 존재하지 않는 오버로드를 호출해
check_function_bodies=on에서 CREATE 시점에 실패한다.

프로덕션 실측(pg_get_functiondef):
  PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price)::numeric AS median_sale_price
저장소 파일에는 ::numeric이 없었다 → 프로덕션 정의에 맞춰 84·99행에 추가.

프로덕션은 이미 이 상태이므로 db push 대상이 아니고 원장과도 무관하다.
로직·값 변경 없음 — 프로덕션과 일치시키는 것이 목적.

+++ supabase/migrations/20260528000003_complex_gap_stats.sql
@@ -81,7 +81,7 @@ AS $$
   WITH sale_stats AS (
     SELECT
       complex_id,
-      PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price) AS median_sale_price,
+      PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price)::numeric AS median_sale_price,
       COUNT(*)::integer AS sale_count
     FROM public.transactions
     WHERE
@@ -96,7 +96,7 @@ AS $$
   jeonse_stats AS (
     SELECT
       complex_id,
-      PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price) AS median_jeonse_price,
+      PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price)::numeric AS median_jeonse_price,
       COUNT(*)::integer AS jeonse_count
     FROM public.transactions
     WHERE
